Bluetooth: Tests: Test restored settings are correct for bonded peers
Test that CCC configuration for bonded peers is correctly restored
across disconnections and reconnection.

That include a case where a peer reconnect to a server previously bonded
with but does not bond again and update CCC configuration. After that,
if that peer disconnect and connect again but this time bond, the
restored CCC configuration should be the one from the last time they
bonded.

Signed-off-by: Théo Battrel <theo.battrel@nordicsemi.no>